ServerPy/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
lxml==5.1.0
protobuf==4.25.3
python-docx==1.1.0
python-dotenv==1.0.1
typing_extensions==4.9.0
